+++ packages/@livestore/expo/src/index.ts
@@ -15,7 +15,7 @@ import { makeMutationEventSchema, MUTATION_LOG_META_TABLE, mutationLogMetaTable 
 import { insertRowPrepared, makeBindValues } from '@livestore/common/sql-queries'
 import { casesHandled, shouldNeverHappen } from '@livestore/utils'
 import { Effect, Option, Queue, Schema, Stream, SubscriptionRef } from '@livestore/utils/effect'
-import * as SQLite from 'expo-sqlite/next'
+import * as SQLite from 'expo-sqlite'
 
 import { makeSynchronousDatabase } from './common.js'
 import type { BootedDevtools } from './devtools.js'