+++ packages/@livestore/utils/src/effect/Effect.ts
@@ -49,7 +49,7 @@ const getThreadName = () => {
     ? globalName
     : // eslint-disable-next-line unicorn/prefer-global-this
       typeof window === 'object'
-      ? 'Browser Main Thread'
+      ? 'Main Thread'
       : 'unknown-thread'
 }